requirements.txt
pytest
flask
python-dotenv
flask-paginate